requirements/dev.txt
-r base.txt
django-debug-toolbar==3.8.1
factory_boy
pytest
pytest-django
